+++ africanus/gridding/wgridder/im2residim.py
@@ -15,7 +15,7 @@ from africanus.util.requirements import requires_optional
 @requires_optional('ducc0.wgridder', ducc_import_error)
 def _residual_internal(uvw, freq, image, vis, freq_bin_idx, freq_bin_counts,
                        cell, weights, flag, celly, epsilon, nthreads,
-                       do_wstacking):
+                       do_wstacking, double_accum):
 
     # adjust for chunking
     # need a copy here if using multiple row chunks
@@ -34,18 +34,21 @@ def _residual_internal(uvw, freq, image, vis, freq_bin_idx, freq_bin_counts,
             mask = flag[:, ind]
         else:
             mask = None
-        residvis = vis[:, ind] - dirty2ms(uvw=uvw, freq=freq[ind],
-                                          dirty=image[i], wgt=None,
-                                          pixsize_x=cell, pixsize_y=celly,
-                                          nu=0, nv=0, epsilon=epsilon,
-                                          nthreads=nthreads, mask=mask,
-                                          do_wstacking=do_wstacking)
+        tvis = vis[:, ind]
+        residvis = tvis - dirty2ms(
+                                uvw=uvw, freq=freq[ind],
+                                dirty=image[i], wgt=None,
+                                pixsize_x=cell, pixsize_y=celly,
+                                nu=0, nv=0, epsilon=epsilon,
+                                nthreads=nthreads, mask=mask,
+                                do_wstacking=do_wstacking)
         residim[0, i] = ms2dirty(uvw=uvw, freq=freq[ind], ms=residvis,
                                  wgt=wgt, npix_x=nx, npix_y=ny,
                                  pixsize_x=cell, pixsize_y=celly,
                                  nu=0, nv=0, epsilon=epsilon,
                                  nthreads=nthreads, mask=mask,
-                                 do_wstacking=do_wstacking)
+                                 do_wstacking=do_wstacking,
+                                 double_precision_accumulation=double_accum)
     return residim
 
 
@@ -54,7 +57,7 @@ def _residual_internal(uvw, freq, image, vis, freq_bin_idx, freq_bin_counts,
 @requires_optional('ducc0.wgridder', ducc_import_error)
 def residual(uvw, freq, image, vis, freq_bin_idx, freq_bin_counts, cell,
              weights=None, flag=None, celly=None, epsilon=1e-5, nthreads=1,
-             do_wstacking=True):
+             do_wstacking=True, double_accum=False):
 
     if celly is None:
         celly = cell
@@ -65,7 +68,8 @@ def residual(uvw, freq, image, vis, freq_bin_idx, freq_bin_counts, cell,
 
     residim = _residual_internal(uvw, freq, image, vis, freq_bin_idx,
                                  freq_bin_counts, cell, weights, flag,
-                                 celly, epsilon, nthreads, do_wstacking)
+                                 celly, epsilon, nthreads, do_wstacking,
+                                 double_accum)
     return residim[0]
 
 
@@ -144,6 +148,9 @@ RESIDUAL_DOCS = DocstringTemplate(
         The number of threads to use. Defaults to one.
     do_wstacking : bool, optional
         Whether to correct for the w-term or not. Defaults to True
+    double_accum : bool, optional
+        If true ducc will accumulate in double precision regardless of
+        the input type.
 
     Returns
     -------